In success case, CGI scripts should return 0, not 1. (#152)
### Issue

When running the included Python server for collectd-web, there are lots
of errors like this in the log (i.e. on stdout);

127.0.0.1 - - [06/Dec/2025 22:02:29] CGI script exit code 1

These are not needed or wanted, since in all three cases the code was
executed successfully.

### Description

Make the CGI script return 0 instead of 1, which is interpreted as an
error.

### Checklist

- [x] I have read the [contributing
guidelines](https://github.com/httpdss/collectd-web/blob/main/README.md#contributing).

Yes - and noticed that the link is broken and should be
https://github.com/httpdss/collectd-web/blob/master/.github/CONTRIBUTING.md
instead.

- [x] My code follows the code style of this project.
- [x] I have performed a self-review of my own code.
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [ ] New and existing unit tests pass locally with my changes.
- [ ] I have updated the documentation accordingly.

### Screenshots (if applicable)
<!-- Add screenshots to illustrate the changes made in the pull request
-->

### Additional Comments
<!-- Add any other context or comments about the pull request here -->